Add files via upload
Added a txt file to remind myself what to work on tomorrow. 

I feel like I am spoending too much time on the notebook and not enough on the documentation. Will be adding multiple graphs via tableau and uploading them to the readme as well as the rst file for the documentation site!